requirements.txt
openai
flask
python-dotenv
Flask-session